adding templates to json chat, now we can add usernames and other dynamic content to the chat
