+++ shared/session.js
@@ -1,0 +1,67 @@
+var passport      = require('passport');
+var LocalStrategy = require('passport-local').Strategy;
+var session       = require('express-session');
+var bodyParser    = require('body-parser');
+var cookieParser  = require('cookie-parser');
+var flash         = require('connect-flash');
+
+module.exports.setup = function (pool) {
+  
+  passport.use(new LocalStrategy(
+    function(username, password, done) {
+        pool.getUser(username, function(user){
+          console.log("user: " + username + " and password: " + password);
+          console.log(user);
+          if (user == undefined) {
+            return done(null, false, { message: 'Incorrect username.' });
+          }
+          if (user.password != password) {
+            return done(null, false, { message: 'Incorrect password.' });
+          }
+          console.log("should be authenticated");
+          return done(null, user);
+        });
+    }
+  ));
+  
+  passport.serializeUser(function(user, done) {
+      done(null, user.username);
+  });
+  
+  passport.deserializeUser(function(username, done) {
+      pool.getUser(username, function( data ){
+        console.log(data);
+        done(null, data);
+      });
+  });
+  
+  function authenticationMiddleware () {
+    return function (req, res, next) {
+  
+      console.log("authenticating...");
+      if (req.isAuthenticated()) {
+        console.log('successful auth');
+        return next()
+      }
+      res.redirect('/login')
+    }
+  }
+  
+  passport.authenticationMiddleware = authenticationMiddleware
+}
+
+module.exports.add_session_to_express = function (app, pool) {
+  app.use(bodyParser.urlencoded({extended: true}));
+  app.use(bodyParser.json());
+  //app.use(cookieParser());
+  app.use(session({
+    store: pool.pgSession,
+    secret: 'ilovescotchscotchyscotchscotch',
+    resave: true,
+    saveUninitialized: true
+  })); // session secret
+  
+  app.use(passport.initialize());
+  app.use(passport.session()); // persistent login sessions
+  app.use(flash()); // use connect-flash for flash messages stored in session
+}
